Refactor: Reorganize Postman translation tests into modules (#4524)
